% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getParamRanges.R
\name{getParamRanges}
\alias{getParamRanges}
\title{Get the parameter ranges used in the Monte Carlo simulations.}
\usage{
getParamRanges()
}
\value{
An environment containing variables and default ranges shown in the function sources.
}
\description{
Generate default parameter ranges for the simulations.
}
\details{
This generates default parameter ranges.  The environment returned
can subsequently be modified by the user.
}

