\name{kb07}
\alias{kb07}
\docType{data}
\title{Kronmueller and Barr, 2007 data}
\description{
  Data described in Kronmueller and Barr, 2007 and referenced
  in Barr et al., 2013
}
\format{
  The format is:
 chr "kb07"
}
\source{
  Barr, Dale J., Levy, Roger, Scheepers, Christoph and Tily, Harry J
  (2013) Random effects structure for confirmatory hypothesis testing:
  Keep it maximal, \emph{Journal of Memory and Language}, \bold{68},
  pp. 255-278
}
\references{
  Kronmueller, Edmundo and Barr, Dale J (2007)
  Perspective-free pragmatics: Broken precedents and the
  recovery-from-preemption hypothesis,  \emph{Journal of Memory and
    Language}, \bold{56}, pp. 436-455
}
\examples{
str(kb07)
}
\keyword{datasets}
