% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mkDf.R
\name{mkCovR}
\alias{mkCovR}
\title{Upper Cholesky factor of covariance}
\usage{
mkCovR(v1, v2, r)
}
\arguments{
\item{v1}{positive numeric, variance of first coordinate}

\item{v2}{positive numeric, variance of second coordinate}

\item{r}{numeric in the range [-1,1], correlation}
}
\value{
The upper triangular Cholesky factor of the covariance matrix
}
\description{
Upper Cholesky factor of a 2 by 2 covariance matrix
}

