% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mkDf.R
\name{mkDf}
\alias{mkDf}
\title{Return a dataframe with simulated data given a set of population parameters}
\usage{
mkDf(nsubj = 24L, nitem = 24L, wsbi = FALSE,
  mcr.params = createParamMx(1L, firstseed = sample.int(1000000L, 1L))[1L, ],
  missMeth = c("random", "none", "randomBig", "bycond", "bysubj",
  "bysubjcond"), rigen = FALSE)
}
\arguments{
\item{nsubj}{number of subjects (default is 24)}

\item{nitem}{number of items (default is 24)}

\item{wsbi}{logical, is design between items (TRUE) or within items (default FALSE).}

\item{mcr.params}{vector of parameters for generation of data (see createParamMx)}

\item{missMeth}{method of generating missing data (default "random")}

\item{rigen}{logical, use a random-intercepts-only generative model (default FALSE)}
}
\value{
a data frame of simulated data
}
\description{
Create simulated data given a row of parameters and the no. of subjects and items
}

