% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rePCA.R
\name{rePCA}
\alias{rePCA}
\title{PCA of random-effects}
\usage{
rePCA(x)
}
\arguments{
\item{x}{a merMod object}
}
\value{
a \code{prcomplist} object
}
\description{
PCA of random-effects variance-covariance estimates
}
\details{
Perform a Principal Components Analysis (PCA) of the random-effects
variance-covariance estimates from a fitted mixed-effects model
}
\author{
Douglas Bates
}

